First hello world app
This will serve as the base for further work on the server. It provides
the "production" Docker image which will build the Go server and serve
it on an Alpine image.
For now, all it does is listen to port 80 and print "Hello world".
The image is tested with goss [0] to have a starting point for a CI/CD
workflow.

[0] https://github.com/aelsabbahy/goss