Add a login form
Adds routes:
- / will redirect to /home if matched exactly. Otherwise, will 404
- /home: Hello world for now. Will be the "Homepage" once logged in
- /login: Login screen

I had to change the port number to 8080 to be able to test locally (go
build .). I can't bind on port 80 locally.

This also changes the production Dockerfile so that it uses a non-root
user (named "mike" for now).

How to test:
- docker build -t hyzual/mike-sierra-sierra .
- run the image
- connect to the docker container's ip on port 8080
- Go to /login
- You should see a (disabled) login screen styled with CSS

It turns out figuring out relative paths for assets and templates was a
bit tricky, you need to get the current working directory (which can
fail).