Allow disabling HTTPS
Since I can't provide on-the-fly valid certificates for the Docker
image, a way to disable HTTPS is needed for it.
A valid certificate and key may still be provided to it through the volume.
Disabling HTTPS will be used to validate the image with dgoss.
It can also be used for dev but generating a cert with mkcert is
preferred.

How to test:
- make build-docker-image
- make dgoss-ci should pass