Have the TypeScript App query the REST endpoint
This adds a GET /api/folders/{id} REST route. For now, it responds a
hardcoded music folder, in order to have the complete "plumbing" between
the frontend app and the server.
This makes sure the TypeScript App can query the protected REST API.
It does send the session cookie with requests, so user can be
authentified by the REST API. It also adds the beginning of a REST
querier interface that will be used for network requests from the
frontend app to the server.

How to test:
- Log in
- In /home, the client app will make a request to /api/folders/0 and
  will log the resulting JSON to the console.

Jest tests will come in a later patch.