Add a "Hello World" REST endpoint
For now it is just to validate that I can create it and have it require
user authentication.
Next step will include building a TypeScript App that can query the
endpoint.
It will be the basis of the client-side app that will query a REST API
to enable discovery of the songs and music folders.