Add a minimal TypeScript app
This contributions adds a minimal (Hello world) Typescript App.
It also introduces all the infrastructure to build, minify and
cache-control assets. Javascript and CSS assets are now output to the
/assets folder via webpack. The folder is emptied by webpack.

This contribution also adds the necessary infrastructure to properly
cache-control assets. Webpack outputs a chunk hash in the filename it
outputs, so that when asset contents change, their filename changes too.
It will allow us to add a Cache-Control header with a long duration to
leverage browser caches for better performances of the frontend.

It also adds the means for the server to resolve the filename of the
assets. Since we don't know at compile-time what the asset name will be,
it must be resolved dynamically. The server reads the manifest.json file
(output by webpack plugin) in the /assets folder and parses its content.
It contains a map from "base name" to "chunk hash name". That way, the
correct GET query is made (with the hashed name) and can be cached.

In order to unit-test the AssetsResolver, I used a virtual filesystem
library that allows me to create an in-memory filesystem with the same
API as the real FS. That way, no temporary files to create and remove.

How to test:
1. access /login
- the page styling should still be there
2. access /home (after logging in)
- The developper console should show a "Hello world" from the TS app
- The stylesheet should be loaded