Add a first-time registration UI
This contribution only adds the web UI. Submitting it leads to a 404 for
now, the handler for the POST request will be handled later.
For now, the page is always accessible, but later, the handler will only
allow access if there is no user in the database.

This contributions also introduces a TemplateExecutor that wraps both
loading the template files and executing them with a given "presenter"
object. This simplifies templated handlers a bit.

This contribution also changes all CSS class-names to prefix them with
"mss-" (for Mike-Sierra-Sierra). It avoids potential naming conflicts
that could arise from external libs.