Handle first-time registration form
All user-related handlers are now grouped together in "user" package.
"server" package no longer depends on "user" package anymore, it is the
other way around.
Common utilities for handlers will be in "server", handlers will be
moved to dedicated package in time.

This refactors LoginHandlers so that they are "erroring handlers" now.
This also introduces "HTTP Error" type. It is an error, but is caught by
server.WrapErrors() which responds according to the error's message and
code. For example, BadRequest error is caught and gives a 400 Bad
Request response.