App navigation is done without reload
The music server is going to be a "single-page app" to avoid reloading
the page on navigation, as it would stop the music playback. So links in
the sidebar will manipulate the history via Javascript and we will add
an event to react on the history and render the correct "page" in
<main>.

In order to facilitate that, the routes will all begin with /app:
/app/home
/app/folders
/app/albums

So the /home route is renamed to /app and the mux router matches all
routes beginning with /app.

This also splits the sidebar template into a dedicated template file.
Same for CSS which has now a dedicated "sidebar" file.