Show a grid of folders with images
The sidebar allows a minimal navigation for now. There are only two
routes:

- Home. It shows nothing by default.
- Folders. It shows a fake list of music folders. For now it is
hardcoded but it must come from the REST API later. As I am still
exploring the requirements for this list, it is easier to just fake them
for now.

Browsing to sub-folders will be handled by a later contribution.

Add a "no-cover" svg illustration

This will be shown when folders or albums or artists don't have a cover
image.

There are two folders: src/ is for Inkscape source SVG and assets/ is
for optimized SVG that will be included in the production application.
References in CSS should always point to assets/ versions.