Use color wheel to adjust color scheme
I did not like the previous dark accent and light accent colors.
I used https://www.canva.com/colors/color-wheel/ to figure out better
colors to match the main color, which I really like.
Now light shades and dark shades have the same Hue but different
Saturation and Luminances as the main color.
Dark accent color is "Analogous" (it is next to the primary color on the
color wheel).
Light accent color is "Complementary" (it is opposite to the primary
color on the color wheel).