Fix dgoss failing
Apparently [0], Docker preserves ownership permissions when you create
volume folders beforehand. VOLUME instruction should be (almost) the
last from the Dockerfile.
This fix allows sqlite to write to the database file. We probably had
the same issue on secrets (though we do not need to write them) and
music volumes, so this changes them also.

[0] https://github.com/docker/compose/issues/3270#issuecomment-363478501