Display Top-level folders
How to test: browse to the "Folders" in the sidebar menu. The "Top-level
folders" API should be queried and for each result, a FolderCover should
be shown.
There is a minimal "Loading ..." state which should be improved later.
The error state is also ugly and should be improved later.
This contribution is more about exploring how to query from a REST API
and show the loading/error/success states. It also leverages neverthrow
ResultAsync monad and lit-html's "until" directive.