Disable write permissions on Actions
This is also an attempt at avoiding to trigger twice the number of CI
jobs on dependabot PRs. The real reason I'm using "push" trigger on
actions is because I don't want to have to open and self-merge PRs, but
if this project (hopefully) becomes maintained by more than myself, I
expect the normal process will be to open and merge PRs.