Browse nested music folders without scanning
How to test:
- using Docker, copy some folders in your /music volume in the dev
container
- Browse the app by clicking on the "Browse By Folder" link
- You should be able to browse nested folders.

Notes: Songs are not yet listed by the frontend. The backend does not
filter which files are "songs", everything that is not a directory is
included.

I don't have the patience to split this massive commit into smaller
parts. This is a squash of a lot of commits that would not make sense to
push as-is. Sorry if you're reviewing this in the future and wonder why
it's all mixed :p.

This includes a reorganization of the go code towards a "Hexagonal
Architecture" model. The principles are:

Domain can never depend on Adapters' code. Adapters call Domain, and
Domain calls interfaces that are implemented by Adapters.
This provides a good separation between side-effects and business domain
code.

There is now only one endpoint to query the contents of a folder. The
"root" music library folder ("/music") is queried the same way, but
providing an empty "path" parameter. It makes it easier to unify the way
to browse nested folders.