Clicking on the "Play" link changes the current song in the MusicPlayer element
How to test:
- add some music files to the docker volume
- when you click on the "Play" link, it changes the "src" attribute of
the <audio> tag in the MusicPlayer element. When you click on the "Play"
control from your browser for said <audio> tag, it plays the song you
clicked on.

AppRoot is now the single element under <body>. It holds the router and
changes the <main> dynamically. It holds all state that needs to be
shared between the different "blocks" of the page:
- the current page that is used to navigate using navigo routes. For now
only the "folders" page is truly handled but later there will be a
"current queue" page.
- the current folder path to navigate in folders.
- the play queue "state". It is a "state" class that is passed as
property to inner components that need to affect the play queue (Play a
song, add to play queue, etc).

Also sorted folder-view components in a dedicated folder to group them
together away from the "root" components.